+++ locustfiles/dataedge_test_1.py
@@ -1,17 +1,17 @@
 """Base Test Design for DataEdge Scenario 1.
 
-For this test we are aiming to download an encrypted file.
+For this test we are aiming to download an unecrypted file.
 The assumption is that the token contains only one file with the correct permissions,
 and we can retrieve the ``file_id`` from the token.
-Scenario 1: Download an encrypted file given a valid token.
+Scenario 1: Download an unencrypted file given a valid token.
 """
 
 import sys
 from ruamel.yaml import YAML
 from locust import HttpLocust, TaskSet, task
-from common import log_format
+from common import log_format, CONFIG_PATH
 
-LOG = log_format('encrypted_dataedge_scenario_5')
+LOG = log_format('dataedge_scenario_1')
 
 
 class APIBehavior(TaskSet):
@@ -24,7 +24,7 @@ class APIBehavior(TaskSet):
                 LOG.error("Data Edge API is not reachable")
                 sys.exit(1)
         yaml = YAML(typ='safe')
-        with open('../stress_tests/enc_dataedge_config.yaml', 'r') as stream:
+        with open(f'{CONFIG_PATH}/dataedge_config.yaml', 'r') as stream:
             self.config = yaml.load(stream)
         if "token" not in self.config['scenario1'] and self.config['scenario1']['token'] is not None:
             LOG.error("Missing Token")
@@ -32,15 +32,12 @@ class APIBehavior(TaskSet):
         else:
             self.token = self.config['scenario1']['token']
             self.file_id = self.config['scenario1']['file_id']
-            self.format = self.config['scenario1']['encrypted']['format']
-            self.key = self.config['scenario1']['encrypted']['key']
-            self.iv = self.config['scenario1']['encrypted']['iv']
             self.ca = self.config['settings']['tls_ca_root_file']
 
     @task
     def get_query(self):
         """Test GET query endpoint."""
-        url = f"/files/{self.file_id}?destinationFormat={self.format}&destinationKey={self.key}&destinationIV={self.iv}"
+        url = f"/files/{self.file_id}?destinationFormat=plain"
         with self.client.get(url,
                              headers={'Authorization': f'Bearer {self.token}'},
                              verify=self.ca,
@@ -50,12 +47,11 @@ class APIBehavior(TaskSet):
 
 
 class APITest(HttpLocust):
-    """Test 5 Encrypted file download via DataEdge API.
+    """Test 1 DataEdge API.
 
     We need an HTTP Locust given the nature of the DataEdge API.
     """
 
     task_set = APIBehavior
-    # longer waiting times due to larger file
-    min_wait = 30000
-    max_wait = 90000
+    min_wait = 5000
+    max_wait = 30000
