+++ locustfiles/dataedge_test_3.py
@@ -9,7 +9,7 @@ Scenario 3: Download multiple unencrypted files given a valid token.
 import sys
 from ruamel.yaml import YAML
 from locust import HttpLocust, TaskSet, task
-from common import log_format
+from common import log_format, CONFIG_PATH
 
 LOG = log_format('dataedge_scenario_3')
 
@@ -24,7 +24,7 @@ class APIBehavior(TaskSet):
                 LOG.error("Data Edge API is not reachable")
                 sys.exit(1)
         yaml = YAML(typ='safe')
-        with open('../stress_tests/dataedge_config.yaml', 'r') as stream:
+        with open(f'{CONFIG_PATH}/dataedge_config.yaml', 'r') as stream:
             self.config = yaml.load(stream)
         if "token" not in self.config['scenario3'] and self.config['scenario3']['token'] is not None:
             LOG.error("Missing Token")
