+++ locustfiles/inbox_test_4.py
@@ -1,14 +1,14 @@
-"""Base Test Design for LocalEGA Inbox Scenario 3.
+"""Base Test Design for LocalEGA Inbox Scenario 4.
 
-For this test we are aiming to upload and rename an encrypted file.
-Scenario 3: Upload an encrypted file and rename file without reconnecting.
+For this test we are aiming to upload, rename and remove an encrypted file.
+Scenario 4: Upload an encrypted file, rename and remove file without reconnecting.
 """
 
 import os
 import paramiko
 from ruamel.yaml import YAML
 from locust import Locust, TaskSet, task
-from common import log_format
+from common import log_format, CONFIG_PATH
 
 LOG = log_format('test_inbox_3')
 
@@ -35,7 +35,7 @@ def open_ssh_connection(hostname, user, key_path, key_pass=None, port=2222):
     return client
 
 
-def sftp_upload_rename(hostname, user, remote_path, new_name, key_path, key_pass=None, port=2222):
+def sftp_upload_rename_remove(hostname, user, remote_path, new_name, key_path, key_pass=None, port=2222):
     """SFTP Client file upload."""
     try:
         k = paramiko.RSAKey.from_private_key_file(key_path, password=key_pass)
@@ -46,6 +46,7 @@ def sftp_upload_rename(hostname, user, remote_path, new_name, key_path, key_pass
         filename, _ = os.path.splitext(remote_path)
         sftp.put(remote_path, f'{filename}'.c4ga)
         sftp.rename(f'{filename}'.c4ga, new_name)
+        sftp.remove(new_name)
         LOG.info(f'file renamed to {new_name}')
     except Exception as e:
         LOG.error(f'Something went wrong {e}')
@@ -61,20 +62,20 @@ class InboxBehavior(TaskSet):
     def setup(self):
         """Test if the inbox is reachable."""
         yaml = YAML(typ='safe')
-        with open('../stress_tests/inbox_config.yaml', 'r') as stream:
+        with open(f'{CONFIG_PATH}/inbox_config.yaml', 'r') as stream:
             self.config = yaml.load(stream)
         # We can reuse the same settings as in scenario2
         self.key_pk = os.path.expanduser(self.config['settings']['user_key'])
         self.user = self.config['settings']['user']
-        self.test_file = os.path.expanduser(self.config['scenario3']['test_file'])
-        self.new_file = os.path.expanduser(self.config['scenario3']['new_file'])
+        self.test_file = os.path.expanduser(self.config['scenario4']['test_file'])
+        self.new_file = os.path.expanduser(self.config['scenario4']['new_file'])
         open_ssh_connection(self.locust.host, self.user, self.key_pk)
 
     @task
     def rename(self):
         """Test upload and rename file."""
-        sftp_upload_rename(self.locust.host, self.user,
-                           self.test_file, self.new_file, self.key_pk)
+        sftp_upload_rename_remove(self.locust.host, self.user,
+                                  self.test_file, self.new_file, self.key_pk)
 
 
 class InboxTest(Locust):
